feat: split trees into separate partitions
