+++ src/modules/tree_algorithms.ixx
@@ -1,0 +1,12 @@
+module;
+
+#include <functional>
+#include <memory>
+#include <vector>
+#include <algorithm>
+
+export module tree_algorithms;
+
+// trees
+export import :BSTree;
+export import :AVLTree;
