+++ src/modules/trees/AVLTree.cpp
@@ -5,91 +5,14 @@ module;
 #include <vector>
 #include <algorithm>
 
-export module tree_algorithms;
+export module tree_algorithms:AVLTree;
+import :BSTree;
 import :nodes;
 
 
-
-export namespace trees {
+namespace trees {
 
 using namespace nodes;
-
-template<typename T, typename Comp = std::less<T>, typename Alloc = std::allocator<T>>
-class BSTree {
-public:
-    virtual void build(const T* items, size_t n) = 0;
-
-    virtual void insert(const T&) = 0;
-
-    virtual size_t count(const T& value) {
-        // TODO: this implementation is too slow!
-        std::vector<TreeNode<T>*> stack;
-
-        if (this->m_root != nullptr) {
-            stack.push_back(this->m_root);
-        }
-
-        size_t answer = 0;
-
-        while (!stack.empty()) {
-            TreeNode<T> *node = stack.back();
-            stack.pop_back();
-
-            if (node->value == value) {
-                ++answer;
-            }
-
-            if (node->left != nullptr) {
-                stack.push_back(node->left);
-            }
-            if (node->right != nullptr) {
-                stack.push_back(node->right);
-            }
-        }
-
-        return answer;
-    }
-
-    virtual bool remove(const T&) = 0;
-
-    virtual void clear() {
-        std::vector<TreeNode<T>*> stack;
-
-        if (this->m_root != nullptr) {
-            stack.push_back(this->m_root);
-        }
-
-        while (!stack.empty()) {
-            TreeNode<T> *node = stack.back();
-            stack.pop_back();
-
-            if (node->left != nullptr) {
-                stack.push_back(node->left);
-            }
-
-            if (node->right != nullptr) {
-                stack.push_back(node->right);
-            }
-
-            this->m_size--;
-            delete node;
-        }
-
-        this->m_root = nullptr;
-    }
-
-    // BSTree<T, Comp, Alloc>::Iterator begin() const
-    // BSTree<T, Comp, Alloc>::Iterator end() const
-
-    virtual ~BSTree() = default;
-
-protected:
-    size_t m_size = 0;
-    TreeNode<T> *m_root = nullptr;
-};
-
-
-
 
 /**
  * Maintains the invariant at each node that the absolute difference in height between subtrees is at most 1.
@@ -97,7 +20,7 @@ protected:
  *
  * Implementation taken from: https://github.com/KadirEmreOto/AVL-Tree
  */
-template<typename T, typename Comp = std::less<T>, typename Alloc = std::allocator<T>>
+export template<typename T, typename Comp = std::less<T>, typename Alloc = std::allocator<T>>
 class AVLTree : public BSTree<T, Comp, Alloc> {
 public:
     void build(const T* items, size_t n) final override {
@@ -209,38 +132,36 @@ public:
 
 private:
     void balance(std::vector<TreeNode<T>**> path) {
-    std::reverse(path.begin(), path.end());
+        std::reverse(path.begin(), path.end());
 
-    for (TreeNode<T>** indirect : path) {
-        AVLTreeNode<T>* node = static_cast<AVLTreeNode<T>*>(*indirect);
+        for (TreeNode<T>** indirect : path) {
+            AVLTreeNode<T>* node = static_cast<AVLTreeNode<T>*>(*indirect);
 
-        AVLTreeNode<T>* left = static_cast<AVLTreeNode<T>*>(node->left);
-        AVLTreeNode<T>* right = static_cast<AVLTreeNode<T>*>(node->right);
+            AVLTreeNode<T>* left = static_cast<AVLTreeNode<T>*>(node->left);
+            AVLTreeNode<T>* right = static_cast<AVLTreeNode<T>*>(node->right);
 
-        node->updateValues();
+            node->updateValues();
 
-        if (node->balanceFactor() >= 2 && left->balanceFactor() >= 0) {
-            // left - left
-            *indirect = node->rightRotate();
-        }
-        else if (node->balanceFactor() >= 2) {
-            // left - right
-            node->left = left->leftRotate();
-            *indirect = node->rightRotate();
-        }
-        else if (node->balanceFactor() <= -2 && right->balanceFactor() <= 0) {
-            // right - right
-            *indirect = node->leftRotate();
-        }
-        else if (node->balanceFactor() <= -2) {
-            // right - left
-            node->right = right->rightRotate();
-            *indirect = node->leftRotate();
+            if (node->balanceFactor() >= 2 && left->balanceFactor() >= 0) {
+                // left - left
+                *indirect = node->rightRotate();
+            }
+            else if (node->balanceFactor() >= 2) {
+                // left - right
+                node->left = left->leftRotate();
+                *indirect = node->rightRotate();
+            }
+            else if (node->balanceFactor() <= -2 && right->balanceFactor() <= 0) {
+                // right - right
+                *indirect = node->leftRotate();
+            }
+            else if (node->balanceFactor() <= -2) {
+                // right - left
+                node->right = right->rightRotate();
+                *indirect = node->leftRotate();
+            }
         }
     }
-}
-
-
 };
 
 }
